+++ app/CurrentState.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App;
+
+enum CurrentState: string
+{
+    case TO_DO = 'TO_DO';
+    case DOING = 'DOING';
+    case DONE = 'DONE';
+}